Added io protection resistors to the schematics. Still need to squeeze them into the logic board.
They may not be really necessary but the atmel application note which is
using the IR2101 bridge drivers has them. So to be sure that nothing
goes wrong we really should add them. On the other hand I had some
problems on the smd prototype of open-bldc which burnt the bridge
drivers it is unlikely but this may be the reason for that. One can
still bridge them when assembling.