Changed default PWM offset to 250.
We realized that we have to sample in pwm on state and use half rail
voltage as transition reference. Removing the caps on the bemf
contitioning signals prevents ramping but you still get some jitter on
the signal. 250 seems to be the right offset from the beginning of the
pwm cycle to avoid hitting the jittering of the bemf signal when pwm
goes high.