Implemented comm detection with linear interpolation.
This means that we use the previous adc value to calculate the real bemf
zero crossing. It is a linearized approach which is not perfect but
improves the stability greatly! The previous border was 3000rpm and the
motor started to rpm swing and stall. Only with this change we reached
4170rpm now at 12V and 4A, on a turingy SK4240-750.