Added spinup routine and cleanups.
Cleanups contain some bugfixes in the BEMF zero crossing detection and
sensor data processing. It is now a real crossing detection and not only
detection if the BEMF is above or under zero. This way we can now detect
if a useful BEMF is being seen and we can start trying to lock on that
signal after spinup.