Fixed a bug preventing RPM > 3090
The variables for calculating new comm time were unsigned that caused an
overflow and erronous detection of direct cutoff boundries. This caused
the commutation frequency to never go above (3090/60)*42=2163Hz and
therefore stalling always at that speed. Next boundry is around 4000rpm
probably because of holdoff and presampling of the bemf signal.