Decreased holdoff (>35000erpm), improved spinup.
erpm stands for electrical field rpm
Another reason for the higher electrical rpm is that now we are not
forcing the bemf detection to lock directly on the bemf crossing signal.
If there was not enough samples to do the crossing detection directly at
the crossing we interpolate to the past to calculate when the crossing
occured.