Fixed PWM pin outputs.
Because of v0.3 hardware changes the GPIOA bank peripheral clock was not
initialized. (it is not used in the LED module anymore) As in the pwm
module not all needed peripheral clocks were initialized the GPIOA bank
clock was untouched. For future reference: each module has to enable
clocks for all peripherals it uses and not only those that are not
initialized by other modules... obviously.