Added automatic dependancy calculation.
Now we are using GCC -MM option to generate a list of .h files that a .c
file depends on. This way we prevent the need of running clean and make
again when we change some .h file that is not listed in the dependancies
like for example config.h.