Added source exist check and moved common objects out.
Using pattern matching targets screws around with gnumake's head. It
does not detect properly when one of the source files is just missing
and says that it can not find the requested target instead of reporting
that the source file is missing. Now added a target that tests if all
provided source files exist and if not aborts with a proper error
message.

Also moved some common source files to COMMON_OBJECTS variable for
better readability and less repetition/mistake probability. :)