Improved commutation frequency aquisition.
If bemf jitters using the last transition as the time we calculate the
frequency from and still using the first transition as the end of the
measurement period. This will force the closed loop to speed up enough
to get rid of most of the jitter. The jitter is caused mostly by bemf
going under 0V on a phase that is still powered. That happens only when
the rpm is not matching the drive power.