Now we are able to compile using libc finally.
Using arm-none-eabi-ld as linker results in an error that libgcc.a can not be found. The reason is that ld does not provide the library search paths. You need to use arm-none-eabi-gcc as a wrapper that provides the library search paths to the linker.

Enabled stripping of debug symbols from libc libgcc and libm as we were already at it.

For some wered reason I had to reenable adc driver again because now it complains about it.